requirements.txt
streamlit==1.45.1
google-genai==1.14.0
PyMuPDF==1.25.5
pdfplumber==0.11.6
python-dotenv==1.1.0
